requirement.txt
# Core libraries
numpy>=1.20.0
opencv-python>=4.5.5
torch>=1.10.0
torchvision>=0.11.0
Pillow>=9.0.0

# YOLO-related
ultralytics>=8.0.0  # For YOLOv8

# Tracking algorithms
filterpy>=1.4.5     # For Kalman filter implementation
scipy>=1.7.0        # For scientific computations
scikit-learn>=1.0.0 # For various algorithms
scikit-image>=0.18.0  # For image processing

# Evaluation metrics
motmetrics>=1.2.0   # For MOT evaluation metrics

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utility
tqdm>=4.62.0
PyYAML>=6.0
